Adapt to updates from landingpage
